API implementation refactoring: adding the FunctionnalBase class
This commit has no impact on applications and demos.

FunctionnalBase defines the basic API for functionnal objects:
 - initCoeffs()
 - eval( x ) const
 - derivative ( ) const
 - operator <<

Classes Functionnal and FunctionnalMapBase inherit from FunctionnalBase, and only different by there constructors and the internal storage type for coefficients.
As a result, the FunctionnalBase API can be extented regardless the type of coefficient storage.